pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
    "clerk-backend-api (>=4.2.0,<5.0.0)",
]
